Remove formatting from files
