+++ packages/oc-docs/src/pages/Request/Request.tsx
@@ -84,7 +84,10 @@ const RequestContent: React.FC<RequestContentProps> = ({
   const body = getHttpBody(item);
   const examples = getRequestExamples(item);
 
-  const { path: pathParams, query: queryParams } = useMemo(() => resolvePathAndQueryParams(params, url), [params, url]);
+  const { path: pathParams, query: queryParams } = useMemo(
+    () => resolvePathAndQueryParams(params, url),
+    [params, url]
+  );
 
   const descHtml = useMemo(() => {
     const content = descriptionContent(item);
@@ -104,10 +107,7 @@ const RequestContent: React.FC<RequestContentProps> = ({
   const scriptChain = useMemo(() => buildScriptChain(collection, ancestry, item), [collection, ancestry, item]);
   const scriptFlow = useMemo(() => getScriptFlow(collection), [collection]);
   const assertions = useMemo(() => collectAssertions(item), [item]);
-  const tests = useMemo(
-    () => collectTests(collection, ancestry, item, scriptFlow),
-    [collection, ancestry, item, scriptFlow]
-  );
+  const tests = useMemo(() => collectTests(collection, ancestry, item, scriptFlow), [collection, ancestry, item, scriptFlow]);
   const testScripts = useMemo(
     () => collectRawTestScripts(collection, ancestry, item, scriptFlow),
     [collection, ancestry, item, scriptFlow]
@@ -137,9 +137,7 @@ const RequestContent: React.FC<RequestContentProps> = ({
       <StyledWrapper className="request" data-testid={testId}>
         <Breadcrumb segments={segments} current={name} onSegmentClick={onBreadcrumbClick} testId="request-breadcrumb" />
 
-        <Heading size="md" style={{ marginTop: '0.875rem' }} testId="request-title">
-          {name}
-        </Heading>
+        <Heading size="md" style={{ marginTop: '0.875rem' }} testId="request-title">{name}</Heading>
 
         <RequestUrlBar method={method} url={url} onTry={onTryClick} style={{ marginTop: '0.9375rem' }} />
 
@@ -199,9 +197,7 @@ const RequestContent: React.FC<RequestContentProps> = ({
           </div>
 
           <div className="request-col-right">
-            <Section label="Code Snippet" testId="request-section-code-snippet">
-              {codeSnippet}
-            </Section>
+            <Section label="Code Snippet" testId="request-section-code-snippet">{codeSnippet}</Section>
           </div>
         </div>
