+++ packages/oc-docs/src/sampleCollection.ts
@@ -118,10 +118,8 @@ items:
     type: "folder"
     items:
       - name: "Live Updates"
-        description: "Hello world"
         type: "websocket"
         url: "{{host}}/ws/updates"
-        docs: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Phasellus felis orci, viverra sed faucibus sed, dapibus id diam. Sed et diam laoreet nulla dignissim ullamcorper. Sed id enim elementum, eleifend quam id, fermentum lacus. Vestibulum sagittis elit libero, sed suscipit urna finibus id. Nam non ante sagittis, viverra nibh sed, semper metus. Donec fringilla ac orci sed aliquet. Proin et varius augue, et porttitor odio. Nulla et hendrerit mi, in viverra dolor. Aliquam erat volutpat. Praesent sagittis dapibus enim sit amet ornare. Curabitur eu enim consectetur lacus rutrum interdum nec et quam. Duis tempor sapien orci, ac lobortis mauris sodales malesuada. Nullam venenatis magna at sem fermentum gravida. Morbi sed orci erat. Sed vel rutrum quam. Aenean vel arcu eget ante lacinia aliquam. \n\nSed vitae enim libero. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Curabitur ullamcorper pulvinar elementum. Donec fringilla ultrices ante, eget tincidunt nulla vestibulum a. Maecenas ac maximus quam, fringilla luctus lacus. Etiam lacinia nunc vitae magna euismod blandit. Cras consequat quam nibh, id aliquam erat convallis in. Fusce accumsan risus sed dolor pretium, nec sagittis metus aliquam. Aliquam erat volutpat. Praesent convallis nulla et diam feugiat bibendum. Vestibulum purus ipsum, pretium sed gravida faucibus, finibus nec nisl. Nunc eleifend justo nec ex vulputate, egestas vehicula nisl ultricies. In ligula tortor, fermentum rhoncus tortor sed, iaculis vehicula tortor. Vivamus ornare tincidunt ultrices. Sed quis nunc id est feugiat pellentesque non eu justo. In vel erat ut quam venenatis fermentum.\nDuis consectetur nibh ut mattis aliquam. Donec suscipit nisl in quam fringilla venenatis. Fusce eu enim tristique, luctus leo ut, ultrices nulla. Vestibulum faucibus consequat dui, et tempus sapien aliquam maximus. Ut lobortis, purus quis luctus imperdiet, arcu erat congue enim, vitae convallis justo ipsum ac tortor. Fusce vehicula elit in sagittis luctus. Donec molestie tristique ex ut volutpat. Aliquam at lorem porta diam eleifend imperdiet. Quisque pretium massa id purus accumsan, eget lacinia lectus dictum. In ultrices magna eu diam commodo cursus. Nunc metus neque, suscipit sed orci vitae, accumsan dictum dolor. Aenean tincidunt sem vel arcu varius, sed sodales libero ullamcorper. Aenean cursus, mi non posuere commodo, orci lacus sollicitudin metus, ac finibus est elit pellentesque tellus. Fusce dui leo, cursus ac magna nec, suscipit finibus erat. Mauris tempus interdum massa, sed sodales nisi pretium vitae.\nVestibulum finibus aliquet quam, mollis cursus justo efficitur nec. Sed ut condimentum lectus. Quisque nisl turpis, pretium ac mi in, posuere mattis lectus. Interdum et malesuada fames ac ante ipsum primis in faucibus. Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos himenaeos. Proin eget suscipit lacus. Integer pulvinar in odio egestas maximus. Quisque egestas quam a lorem volutpat feugiat. Aenean vel arcu augue. Nulla sit amet enim sem. Curabitur pellentesque porttitor mi vitae dapibus. Nam at metus dui. Suspendisse libero risus, porta in tincidunt egestas, rhoncus a urna. Proin ut ante tempus, mollis tellus quis, eleifend urna.\nCurabitur magna ex, iaculis aliquam libero ac, molestie dictum est. In magna turpis, finibus nec neque ut, feugiat tincidunt augue. Cras auctor at eros ut ullamcorper. Nulla iaculis feugiat orci et placerat. Donec euismod urna id semper sollicitudin. Quisque suscipit hendrerit lectus id blandit. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas. Vestibulum bibendum purus nunc, sed lacinia magna posuere non. Integer eu imperdiet nibh. Vivamus vestibulum elementum sapien, nec gravida purus auctor ut."
       - name: "GraphQL API"
         type: "graphql"
         url: "{{host}}/graphql"
